Move Seed XLSX logic into a standalone service
