+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/seed/SeedService.kt
@@ -1,14 +1,11 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.seed
 
 import com.github.doyaaaaaken.kotlincsv.dsl.csvReader
-import org.jetbrains.kotlinx.dataframe.DataFrame
-import org.jetbrains.kotlinx.dataframe.io.readExcel
 import org.springframework.context.ApplicationContext
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Service
 import org.springframework.transaction.support.TransactionTemplate
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SeedFileType
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.SeedFromExcelFileType
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.config.SeedConfig
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.ApStaffUsersSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.ApprovedPremisesBookingCancelSeedJob
@@ -24,8 +21,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1OutOfServi
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1PlanSpacePlanningDryRunSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1RemoveAssessmentDetailsSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1SeedPremisesFromCsvJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1SeedPremisesFromSiteSurveyXlsxJob
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1SeedRoomsFromSiteSurveyXlsxJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1UpdateEventNumberSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1UpdateSpaceBookingSeedJob
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.seed.cas1.Cas1UsersSeedJob
@@ -41,7 +36,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.util.findRootCause
 import java.nio.file.Path
 import java.time.LocalDateTime
 import java.time.temporal.ChronoUnit
-import java.util.UUID
 import kotlin.io.path.absolutePathString
 import kotlin.reflect.KClass
 
@@ -56,40 +50,6 @@ class SeedService(
   fun seedDataAsync(seedFileType: SeedFileType, filename: String) = seedData(seedFileType, filename)
 
   fun seedData(seedFileType: SeedFileType, filename: String) = seedData(seedFileType, filename) { "${seedConfig.filePrefix}/$filename" }
-
-  fun seedExcelData(excelSeedFileType: SeedFromExcelFileType, premisesId: UUID, filename: String) =
-    seedExcelData(excelSeedFileType, premisesId, filename) { "${seedConfig.filePrefix}/$filename" }
-
-  @SuppressWarnings("TooGenericExceptionThrown")
-  fun seedExcelData(excelSeedFileType: SeedFromExcelFileType, premisesId: UUID, filename: String, resolveXlsxPath: ExcelSeedJob.() -> String) {
-    seedLogger.info("Starting seed request: $excelSeedFileType - $filename")
-
-    try {
-      if (filename.contains("/") || filename.contains("\\")) {
-        throw RuntimeException("Filename must be just the filename of a .xlsx file in the /seed directory, e.g. for /seed/upload.xlsx, just `upload` should be supplied")
-      }
-
-      val jobAndSeed = when (excelSeedFileType) {
-        SeedFromExcelFileType.CAS1_IMPORT_SITE_SURVEY_ROOMS -> Pair(
-          getBean(Cas1SeedRoomsFromSiteSurveyXlsxJob::class),
-          "Sheet3",
-        )
-        SeedFromExcelFileType.CAS1_IMPORT_SITE_SURVEY_PREMISES -> Pair(
-          getBean(Cas1SeedPremisesFromSiteSurveyXlsxJob::class),
-          "Sheet2",
-        )
-      }
-
-      val seedStarted = LocalDateTime.now()
-
-      transactionTemplate.executeWithoutResult { processExcelJob(jobAndSeed.first, premisesId, jobAndSeed.second, resolveXlsxPath) }
-
-      val timeTaken = ChronoUnit.MILLIS.between(seedStarted, LocalDateTime.now())
-      seedLogger.info("Excel seed request complete. Took $timeTaken millis")
-    } catch (exception: Exception) {
-      seedLogger.error("Unable to complete Excel seed job", exception)
-    }
-  }
 
   @SuppressWarnings("CyclomaticComplexMethod", "TooGenericExceptionThrown")
   fun seedData(seedFileType: SeedFileType, filename: String, resolveCsvPath: SeedJob<*>.() -> String) {
@@ -162,17 +122,6 @@ class SeedService(
     job.postSeed()
 
     return rowsProcessed
-  }
-
-  @Suppress("TooGenericExceptionThrown")
-  private fun processExcelJob(job: ExcelSeedJob, premisesId: UUID, sheetName: String, resolveXlsxPath: ExcelSeedJob.() -> String) {
-    seedLogger.info("Processing XLSX file ${Path.of(job.resolveXlsxPath()).absolutePathString()}")
-    try {
-      val dataFrame = DataFrame.readExcel(job.resolveXlsxPath(), sheetName)
-      job.processDataFrame(dataFrame, premisesId)
-    } catch (exception: Exception) {
-      throw RuntimeException("Unable to process XLSX file", exception)
-    }
   }
 
   private fun <T> processCsv(job: SeedJob<T>, resolveCsvPath: SeedJob<T>.() -> String): Int {
